app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
neo4j==5.22.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1


